move some files and write a simple unit test to cover the send app telemetry initializer part
