+++ src/electron/device-connect-view/device-connect-view-initializer.ts
@@ -22,10 +22,10 @@ import { telemetryAppTitle } from '../../content/strings/application';
 import { ElectronAppDataAdapter } from '../adapters/electron-app-data-adapter';
 import { ElectronStorageAdapter } from '../adapters/electron-storage-adapter';
 import { RiggedFeatureFlagChecker } from '../common/rigged-feature-flag-checker';
-import { sendAppInitializedTelemetryEvent } from '../common/send-app-initialized-telemetry';
 import { DeviceConnectActionCreator } from '../flux/action-creator/device-connect-action-creator';
 import { ElectronLink } from './components/electron-link';
 import { DeviceConnectViewRenderer } from './device-connect-view-renderer';
+import { sendAppInitializedTelemetryEvent } from './send-app-initialized-telemetry';
 
 initializeFabricIcons();
 
@@ -77,5 +77,6 @@ getPersistedData(indexedDBInstance, indexedDBDataKeysToFetch).then((persistedDat
     const renderer = new DeviceConnectViewRenderer(ReactDOM.render, document, props);
     renderer.render();
 
+    // send app initialized telemetry - depending on user consent
     sendAppInitializedTelemetryEvent(telemetryEventHandler);
 });
